requirements.txt
msal==1.25.0
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.0
openai
markitdown

# Testing dependencies
pytest>=7.4.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0